Fix warnings level 3

+++ src/app/camera.cpp
@@ -121,21 +121,21 @@ void camera::update(const rh::ctx& ctx)
         movements_[i].step = new_step;
         if (rosy_math::is_equal(new_step.current_state.velocity, 0.0))
         {
-            to_remove = i;
+            to_remove = static_cast<int>(i);
         }
 	    switch (mv.dir)
 	    {
 		case movement::horizontal:
             updated = true;
-            vel[0] = mv.step.current_state.position;
+            vel[0] = static_cast<float>(mv.step.current_state.position);
 			break;
 		case movement::vertical:
             updated = true;
-            vel[1] = mv.step.current_state.position;
+            vel[1] = static_cast<float>(mv.step.current_state.position);
 			break;
         case movement::direction::depth:
             updated = true;
-            vel[2] = mv.step.current_state.position;
+            vel[2] = static_cast<float>(mv.step.current_state.position);
             break;
 	    }
     }
